+++ sdk/aqueduct/artifacts/generic_artifact.py
@@ -2,23 +2,26 @@ from __future__ import annotations
 
 import json
 import uuid
-from typing import Any, Dict, Optional, Union
+from typing import Any, Dict, List, Optional, Union
 
-from aqueduct.artifacts import preview as artifact_utils
-from aqueduct.artifacts.base_artifact import BaseArtifact
-from aqueduct.artifacts._create import create_metric_or_check_artifact
-from aqueduct.constants.enums import ArtifactType, ExecutionMode, ExecutionStatus
-from aqueduct.error import ArtifactNeverComputedException
-from aqueduct.models.dag import DAG
-from aqueduct.utils.utils import format_header_for_print, generate_uuid
 from aqueduct.artifacts import bool_artifact, numeric_artifact
+from aqueduct.artifacts import preview as artifact_utils
+from aqueduct.artifacts import system_metric
+from aqueduct.artifacts._create import create_metric_or_check_artifact
+from aqueduct.artifacts.base_artifact import BaseArtifact
+from aqueduct.constants.enums import ArtifactType, ExecutionMode, ExecutionStatus
 from aqueduct.constants.metrics import SYSTEM_METRICS_INFO
-from aqueduct.models.operators import Operator, OperatorSpec, SystemMetricSpec
+from aqueduct.error import ArtifactNeverComputedException
 from aqueduct.models.artifact import ArtifactMetadata
+from aqueduct.models.dag import DAG
+from aqueduct.models.operators import Operator, OperatorSpec, SystemMetricSpec
 from aqueduct.utils.naming import default_artifact_name_from_op_name
+from aqueduct.utils.utils import format_header_for_print, generate_uuid
+
+from aqueduct import globals
 
 
-class GenericArtifact(BaseArtifact):
+class GenericArtifact(BaseArtifact, system_metric.SystemMetricMixin):
     """This class represents a generic artifact within the flow's DAG.
 
     Currently, a generic artifact can be any artifact other than table, numeric, bool, or parameter
@@ -96,7 +99,7 @@ class GenericArtifact(BaseArtifact):
         ]
         print(format_header_for_print(f"'{input_operator.name}' {self.type()} Artifact"))
         print(json.dumps(readable_dict, sort_keys=False, indent=4))
-    
+
     def system_metric(
         self, metric_name: str, lazy: bool = False
     ) -> numeric_artifact.NumericArtifact:
@@ -114,72 +117,5 @@ class GenericArtifact(BaseArtifact):
         """
         if globals.__GLOBAL_CONFIG__.lazy:
             lazy = True
-        execution_mode = ExecutionMode.EAGER if not lazy else ExecutionMode.LAZY
 
-        operator = self._dag.must_get_operator(with_output_artifact_id=self._artifact_id)
-        system_metric_description, system_metric_unit = SYSTEM_METRICS_INFO[metric_name]
-        system_metric_name = "%s %s(%s) metric" % (operator.name, metric_name, system_metric_unit)
-        op_spec = OperatorSpec(system_metric=SystemMetricSpec(metric_name=metric_name))
-        new_artifact = self._apply_operator_to_table(
-            op_spec,
-            system_metric_name,
-            system_metric_description,
-            output_artifact_type_hint=ArtifactType.NUMERIC,
-            execution_mode=execution_mode,
-        )
-
-        assert isinstance(new_artifact, numeric_artifact.NumericArtifact)
-
-        return new_artifact
-
-    def _apply_operator_to_table(
-        self,
-        op_spec: OperatorSpec,
-        op_name: str,
-        op_description: str,
-        output_artifact_type_hint: ArtifactType,
-        execution_mode: ExecutionMode = ExecutionMode.EAGER,
-    ) -> Union[numeric_artifact.NumericArtifact, bool_artifact.BoolArtifact]:
-        assert (
-            output_artifact_type_hint == ArtifactType.NUMERIC
-            or output_artifact_type_hint == ArtifactType.BOOL
-        )
-
-        operator_id = generate_uuid()
-        output_artifact_id = generate_uuid()
-        artifact_name = default_artifact_name_from_op_name(op_name)
-
-        create_metric_or_check_artifact(
-            dag=self._dag,
-            op=Operator(
-                id=operator_id,
-                name=op_name,
-                description=op_description,
-                spec=op_spec,
-                inputs=[self._artifact_id],
-                outputs=[output_artifact_id],
-            ),
-            output_artifacts=[
-                ArtifactMetadata(
-                    id=output_artifact_id,
-                    name=artifact_name,
-                    type=output_artifact_type_hint,
-                    explicitly_named=False,
-                )
-            ],
-        )
-
-        if execution_mode == ExecutionMode.EAGER:
-            # Issue preview request since this is an eager execution.
-            artifact = artifact_utils.preview_artifact(self._dag, output_artifact_id)
-
-            assert isinstance(artifact, numeric_artifact.NumericArtifact) or isinstance(
-                artifact, bool_artifact.BoolArtifact
-            )
-            return artifact
-        else:
-            # We are in lazy mode.
-            if output_artifact_type_hint == ArtifactType.NUMERIC:
-                return numeric_artifact.NumericArtifact(self._dag, output_artifact_id)
-            else:
-                return bool_artifact.BoolArtifact(self._dag, output_artifact_id)
+        return self.system_metric_helper(self._dag, self._artifact_id, metric_name, lazy)
